editAddin {rhandsontable}	R Documentation

Edit a Data Frame.

Description

Interactively edit a data.frame or data.table. The resulting code will be emitted as a call to reload the data from a temp RDS file.

Usage

editAddin()

Details

This addin can be used to interactively edit. The intended way to use this is as follows:

1. Highlight a symbol naming a data.frame or data.table in your R session, e.g. mtcars. 2. Execute this addin, to interactively edit it.

When you're done, the code performing this operation will be emitted at the cursor position.

This function borrows heavily from https://github.com/rstudio/addinexamples/blob/master/R/subsetAddin.R

[Package rhandsontable version 0.3.5 Index]