hot_cols {rhandsontable}	R Documentation

Handsontable widget

Description

Configure multiple columns.

Usage

hot_cols(hot, colWidths = NULL, columnSorting = NULL,
  manualColumnMove = NULL, manualColumnResize = NULL,
  fixedColumnsLeft = NULL, ...)

Arguments

hot
rhandsontable object

colWidths
a scalar or numeric vector of column widths

columnSorting
logical enabling row sorting. Sorting only alters the table presentation and the original dataset row order is maintained. The sorting will be done when a user click on column name

manualColumnMove
logical enabling column drag-and-drop reordering

manualColumnResize
logical enabline column width resizing

fixedColumnsLeft
a scalar indicating the number of columns to freeze on the left

...
passed to hot_col

See Also

hot_col, hot_rows, hot_cell

Examples

library(rhandsontable)
DF = data.frame(val = 1:10, bool = TRUE, big = LETTERS[1:10],
                small = letters[1:10],
                dt = seq(from = Sys.Date(), by = "days", length.out = 10),
                stringsAsFactors = FALSE)

rhandsontable(DF) %>%
  hot_cols(columnSorting = TRUE)

[Package rhandsontable version 0.3.5 Index]