hot_heatmap {rhandsontable}	R Documentation

Handsontable widget

Description

Add heatmap to table. See Heatmaps for values in a column for details.

Usage

hot_heatmap(hot, cols, color_scale = c("#ED6D47", "#17F556"),
  renderer = NULL)

Arguments

hot
rhandsontable object

cols
numeric vector of columns to include in the heatmap. If missing all columns are used.

color_scale
character vector that includes the lower and upper colors

renderer
character defining a Javascript function to be used to determine the cell colors. If missing, rhandsontable:::renderer_heatmap is used.

Examples

MAT = matrix(rnorm(50), nrow = 10, dimnames = list(LETTERS[1:10],
             letters[1:5]))

rhandsontable(MAT) %>%
 hot_heatmap()

[Package rhandsontable version 0.3.5 Index]