rHandsontableOutput {rhandsontable}	R Documentation

Handsontable widget

Description

Shiny bindings for rhandsontable

Usage

rHandsontableOutput(outputId, width = "100%", height = "100%")

Arguments

outputId
output variable to read from

width, height
must be a valid CSS unit in pixels or a number, which will be coerced to a string and have "px" appended.

See Also

renderRHandsontable

[Package rhandsontable version 0.3.5 Index]